Also support remote-objects.yaml
We only applied `GREMLIN_REMOTE_HOSTS` to the `remote.yaml`, but we
should also use it for the `remote-objects.yaml` which can for example
be used to create remote traversals instead of sending traversals as
strings to the server.

This change was initially made by @phreed in PR #88
https://github.com/JanusGraph/janusgraph-docker/pull/88

who unfortunately didn't come back to it to update the PR to only update
the latest version (currently 0.6).

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>