Add an image for 1.0.0 RC1
This required some changes to the Bash scripts:
* Parsing versions from GitHub API needs to include pre-release versions
* Pre-release versions should only be tagged with their full version
* Pre-release versions should be ignored for `latest`

I did not add logic to detect if there is a release candidate for a
minor version for which we already have a stable version. In that case
we would overwrite the stable version with the release candidate if it
is newer.
Example: We already have 1.0.0 and then publish 1.0.1-rc1. This would
change the version in the 1.0 directory to 1.0.1-rc1 which we probably
don't want.
I think it is OK to not handle that for now as we usually don't publish
release candidates for a simple patch release (like 1.0.1), but only
for at least a minor version bump.
If we however run into this problem, then we should still catch it when
we update the image as that is a manual process.

As Java 11 is recommended for versions 1.0 and higher, we also use an
11-jre image as a base for these versions.

I've removed the Gryo MessageSerializers as they are no longer supported
in 1.0.0.
log4-server.properties needs to be removed as 1.0.0 uses Log4j2 which
has a different config format: https://docs.janusgraph.org/master/changelog/#upgrade-of-log4j-to-version-2
I however don't  think that we have to add one here and should instead
use the default config from the distribution build.

Fixes #111

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>